Extract folder browser view module
